# docs/api/transaction-service.openapi.json
{
  "components" : {
    "schemas" : {
...
      "CustomerStatementGenerateRequest" : {
        "properties" : {
          "externalAccountId" : {
            "minLength" : 1,
            "type" : "string"
          },
          "yearMonth" : {
            "minLength" : 1,
            "pattern" : "\\d{4}-(0[1-9]|1[0-2])",
            "type" : "string"
          }
        },
        "required" : [ "externalAccountId", "yearMonth" ],
        "type" : "object"
      },
      "CustomerStatementLineResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "currency" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "effectiveDate" : {
            "format" : "date",
            "type" : "string"
          },
          "journalId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "lineId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "lineSequence" : {
            "format" : "int32",
            "type" : "integer"
          },
          "runningBalance" : {
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "CustomerStatementResponse" : {
        "properties" : {
          "closingBalance" : {
            "type" : "number"
          },
          "currency" : {
            "type" : "string"
          },
          "externalAccountId" : {
            "type" : "string"
          },
          "generatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "lines" : {
            "items" : {
              "$ref" : "#/components/schemas/CustomerStatementLineResponse"
            },
            "type" : "array"
          },
          "openingBalance" : {
            "type" : "number"
          },
          "periodEnd" : {
            "format" : "date",
            "type" : "string"
          },
          "periodStart" : {
            "format" : "date",
            "type" : "string"
          },
          "statementId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "statementVersion" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "Problem" : {
        "description" : "RFC 9457 Problem Details. error, message, path and timestamp repeat title, detail and instance for clients written before the RFC format.",
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "error" : {
            "type" : "string"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "transactionId" : {
            "type" : "string"
          },
          "type" : {
            "example" : "urn:financial:problem:insufficient-funds",
            "format" : "uri",
            "type" : "string"
          },
          "validationErrors" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "description" : "Field name to message, for validation failures",
            "type" : "object"
          }
        },
        "required" : [ "type", "title", "status" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Transfers, deposits/withdrawals, scheduled transfers, ledger, disputes, risk and Outcome Protection. Errors use RFC 9457 Problem Details.",
    "title" : "Transaction Service API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/ledger/statements" : {
      "get" : {
        "operationId" : "list_2",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CustomerStatementResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error, as RFC 9457 Problem Details"
          }
        },
        "tags" : [ "customer-statement-controller" ]
      },
      "post" : {
        "operationId" : "generate",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CustomerStatementGenerateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerStatementResponse"
                }
              }
            },
            "description" : "OK"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error, as RFC 9457 Problem Details"
          }
        },
        "tags" : [ "customer-statement-controller" ]
      }
    },
    "/api/ledger/statements/{statementId}" : {
      "get" : {
        "operationId" : "get_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "statementId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerStatementResponse"
                }
              }
            },
            "description" : "OK"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error, as RFC 9457 Problem Details"
          }
        },
        "tags" : [ "customer-statement-controller" ]
      }
    },
    "/api/ledger/statements/{statementId}/csv" : {
      "get" : {
        "operationId" : "exportCsv",
        "parameters" : [ {
          "in" : "path",
          "name" : "statementId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "text/csv" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error, as RFC 9457 Problem Details"
          }
        },
        "tags" : [ "customer-statement-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Through the gateway",
    "url" : "/transaction-api"
  } ]
}